Rename package `diary` to `home`
