+++ app/src/main/java/com/chaeniiz/weatherdiary/presentation/home/HomeRecyclerAdapter.kt
@@ -1,4 +1,4 @@
-package com.chaeniiz.weatherdiary.presentation.diary
+package com.chaeniiz.weatherdiary.presentation.home
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -7,15 +7,15 @@ import com.chaeniiz.entity.entities.Diary
 import com.chaeniiz.weatherdiary.R
 import org.jetbrains.anko.onClick
 
-class DiaryRecyclerAdapter(
+class HomeRecyclerAdapter(
     private val diaries: List<Diary>,
     private val onItemClicked: (id: Int) -> Unit
-) : RecyclerView.Adapter<DiaryViewHolder>() {
+) : RecyclerView.Adapter<HomeViewHolder>() {
 
     override fun getItemCount(): Int = diaries.size
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): DiaryViewHolder =
-        DiaryViewHolder(
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): HomeViewHolder =
+        HomeViewHolder(
             LayoutInflater.from(parent.context).inflate(R.layout.item_diary, parent, false)
         ).apply {
             itemView.onClick {
@@ -23,7 +23,7 @@ class DiaryRecyclerAdapter(
             }
         }
 
-    override fun onBindViewHolder(holder: DiaryViewHolder, position: Int) {
+    override fun onBindViewHolder(holder: HomeViewHolder, position: Int) {
         holder.bind(diaries[position])
     }
 }
