+++ app/src/main/res/layout/activity_home.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.diary.DiaryActivity">
+    tools:context=".presentation.home.HomeActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/diaryRecyclerView"
